+++ fidget-wgpu/src/effects/mod.rs
@@ -10,9 +10,10 @@
 //! - Apply shading to a [`PackedVoxel`] buffer, producing an RGBA image buffer
 
 use crate::{
+    Gpu,
     buf::{
-        BufferSizeError, ImageBuffer, ImageReadBuffer, MappedImage, buffer_ro,
-        buffer_rw, buffer_uniform,
+        BufferSizeError, ImageBuffer, ImageReadBuffer, buffer_ro, buffer_rw,
+        buffer_uniform,
     },
     tag,
     voxel::GeomBufferTag,
@@ -22,19 +23,22 @@ use zerocopy::{FromBytes, Immutable, IntoBytes, KnownLayout};
 
 /// WGPU context for applying various effects
 pub struct Context {
-    device: wgpu::Device,
-    queue: wgpu::Queue,
+    gpu: Gpu,
 
     merge_bind_group_layout: wgpu::BindGroupLayout,
     merge_pipeline: wgpu::ComputePipeline,
 
     shade_bind_group_layout: wgpu::BindGroupLayout,
     shade_pipeline: wgpu::ComputePipeline,
+
+    ssao_ctx: SsaoContext,
 }
 
 const COMMON_SHADER: &str = include_str!("shaders/common.wgsl");
 const MERGE_SHADER: &str = include_str!("shaders/merge.wgsl");
 const SHADE_SHADER: &str = include_str!("shaders/shade.wgsl");
+const SSAO_SHADER: &str = include_str!("shaders/ssao.wgsl");
+const BLUR_SHADER: &str = include_str!("shaders/blur.wgsl");
 
 fn merge_shader() -> String {
     MERGE_SHADER.to_owned() + COMMON_SHADER + crate::COMMON_SHADER
@@ -42,6 +46,14 @@ fn merge_shader() -> String {
 
 fn shade_shader() -> String {
     SHADE_SHADER.to_owned() + COMMON_SHADER + crate::COMMON_SHADER
+}
+
+fn ssao_shader() -> String {
+    SSAO_SHADER.to_owned() + COMMON_SHADER + crate::COMMON_SHADER
+}
+
+fn blur_shader() -> String {
+    BLUR_SHADER.to_owned() + COMMON_SHADER + crate::COMMON_SHADER
 }
 
 /// Packed voxel structure used on the GPU
@@ -90,7 +102,9 @@ struct MergeConfig {
 struct ShadeConfig {
     /// Image size, in pixels
     image_size: [u32; 3],
-    _pad: u32,
+
+    /// Flag to determine whether the SSAO buffer is valid (0 / 1)
+    has_ssao: u32,
 }
 
 tag!(MergeVoxelBufferTag, PackedVoxel, STORAGE | COPY_SRC);
@@ -113,12 +127,20 @@ pub struct ShadeBuffers {
     out: ImageBuffer<ShadedImageTag>,
 }
 
+impl ShadeBuffers {
+    /// Returns a reference to the output buffer
+    pub fn output(&self) -> &ImageBuffer<ShadedImageTag> {
+        &self.out
+    }
+}
+
 /// Error returned when submitting a merge operation
 #[derive(Debug, thiserror::Error)]
 pub enum MergeError {
     /// Image sizes in the slice are not consistent
     #[error(transparent)]
     ImageSizeMismatch(#[from] ImageSizeMismatch),
+
     /// An error occurred while resizing the output buffer
     #[error(transparent)]
     OutputSize(BufferSizeError),
@@ -127,6 +149,14 @@ pub enum MergeError {
 /// Error returned when submitting a shade operation
 #[derive(Debug, thiserror::Error)]
 pub enum ShadeError {
+    /// An error occurred while resizing the output buffer
+    #[error(transparent)]
+    OutputSize(BufferSizeError),
+}
+
+/// Error returned when submitting an SSAO operation
+#[derive(Debug, thiserror::Error)]
+pub enum SsaoError {
     /// An error occurred while resizing the output buffer
     #[error(transparent)]
     OutputSize(BufferSizeError),
@@ -146,9 +176,9 @@ pub struct ImageSizeMismatch {
 
 impl Context {
     /// Builds a new context for applying effects
-    pub fn new(device: wgpu::Device, queue: wgpu::Queue) -> Self {
-        let merge_bind_group_layout =
-            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+    pub fn new(gpu: &Gpu) -> Self {
+        let merge_bind_group_layout = gpu.device.create_bind_group_layout(
+            &wgpu::BindGroupLayoutDescriptor {
                 label: None,
                 entries: &[
                     buffer_uniform(0),
@@ -161,67 +191,78 @@ impl Context {
                     buffer_ro(7), // image6
                     buffer_rw(8), // out
                 ],
-            });
+            },
+        );
         let shader_code = merge_shader();
-        let pipeline_layout =
-            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+        let pipeline_layout = gpu.device.create_pipeline_layout(
+            &wgpu::PipelineLayoutDescriptor {
                 label: Some("effects merge pipeline"),
                 bind_group_layouts: &[Some(&merge_bind_group_layout)],
                 immediate_size: 0u32,
-            });
+            },
+        );
         let shader_module =
-            device.create_shader_module(wgpu::ShaderModuleDescriptor {
-                label: Some("effects merge shader module"),
-                source: wgpu::ShaderSource::Wgsl(shader_code.into()),
-            });
-        let merge_pipeline =
-            device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
+            gpu.device
+                .create_shader_module(wgpu::ShaderModuleDescriptor {
+                    label: Some("effects merge shader module"),
+                    source: wgpu::ShaderSource::Wgsl(shader_code.into()),
+                });
+        let merge_pipeline = gpu.device.create_compute_pipeline(
+            &wgpu::ComputePipelineDescriptor {
                 label: Some("effects merge compute pipeline"),
                 layout: Some(&pipeline_layout),
                 module: &shader_module,
                 entry_point: Some("merge_main"),
                 compilation_options: Default::default(),
                 cache: None,
-            });
+            },
+        );
 
-        let shade_bind_group_layout =
-            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+        let shade_bind_group_layout = gpu.device.create_bind_group_layout(
+            &wgpu::BindGroupLayoutDescriptor {
                 label: None,
                 entries: &[
                     buffer_uniform(0),
                     buffer_ro(1), // image
-                    buffer_rw(2), // out
+                    buffer_ro(2), // ssao occlusion
+                    buffer_rw(3), // out
                 ],
-            });
+            },
+        );
         let shader_code = shade_shader();
-        let pipeline_layout =
-            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+        let pipeline_layout = gpu.device.create_pipeline_layout(
+            &wgpu::PipelineLayoutDescriptor {
                 label: Some("effects shade pipeline"),
                 bind_group_layouts: &[Some(&shade_bind_group_layout)],
                 immediate_size: 0u32,
-            });
+            },
+        );
         let shader_module =
-            device.create_shader_module(wgpu::ShaderModuleDescriptor {
-                label: Some("effects shade shader module"),
-                source: wgpu::ShaderSource::Wgsl(shader_code.into()),
-            });
-        let shade_pipeline =
-            device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
+            gpu.device
+                .create_shader_module(wgpu::ShaderModuleDescriptor {
+                    label: Some("effects shade shader module"),
+                    source: wgpu::ShaderSource::Wgsl(shader_code.into()),
+                });
+        let shade_pipeline = gpu.device.create_compute_pipeline(
+            &wgpu::ComputePipelineDescriptor {
                 label: Some("effects shade compute pipeline"),
                 layout: Some(&pipeline_layout),
                 module: &shader_module,
                 entry_point: Some("shade_main"),
                 compilation_options: Default::default(),
                 cache: None,
-            });
+            },
+        );
+
+        let ssao_ctx = SsaoContext::new(&gpu.device);
 
         Self {
-            device,
-            queue,
+            gpu: gpu.clone(),
             merge_bind_group_layout,
             merge_pipeline,
             shade_bind_group_layout,
             shade_pipeline,
+            ssao_ctx,
         }
     }
 
@@ -230,14 +271,14 @@ impl Context {
         &self,
         image_size: VoxelSize,
     ) -> Result<MergeBuffers, BufferSizeError> {
-        let config = self.device.create_buffer(&wgpu::BufferDescriptor {
+        let config = self.gpu.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("config"),
             size: std::mem::size_of::<MergeConfig>() as u64,
             usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
             mapped_at_creation: false,
         });
         let out = ImageBuffer::new(
-            &self.device,
+            &self.gpu.device,
             "merge output".to_owned(),
             ImageSize::new(image_size.width(), image_size.height()),
         )?;
@@ -253,18 +294,57 @@ impl Context {
         &self,
         image_size: ImageSize,
     ) -> Result<ShadeBuffers, BufferSizeError> {
-        let config = self.device.create_buffer(&wgpu::BufferDescriptor {
+        let config = self.gpu.device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("shade config"),
             size: std::mem::size_of::<ShadeConfig>() as u64,
             usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
             mapped_at_creation: false,
         });
         let out = ImageBuffer::new(
-            &self.device,
+            &self.gpu.device,
             "shade output".to_owned(),
             ImageSize::new(image_size.width(), image_size.height()),
         )?;
         Ok(ShadeBuffers { config, out })
+    }
+
+    /// Builds a new set of [`SsaoBuffers`] for the given image size
+    pub fn ssao_buffers(
+        &self,
+        image_size: VoxelSize,
+    ) -> Result<SsaoBuffers, BufferSizeError> {
+        let ssao_config =
+            self.gpu.device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some("ssao config"),
+                size: std::mem::size_of::<SsaoConfig>() as u64,
+                usage: wgpu::BufferUsages::UNIFORM
+                    | wgpu::BufferUsages::COPY_DST,
+                mapped_at_creation: false,
+            });
+        let raw_occlusion = ImageBuffer::new(
+            &self.gpu.device,
+            "ssao raw occlusion".to_owned(),
+            ImageSize::new(image_size.width(), image_size.height()),
+        )?;
+        let blur_config =
+            self.gpu.device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some("blur config"),
+                size: std::mem::size_of::<BlurConfig>() as u64,
+                usage: wgpu::BufferUsages::UNIFORM
+                    | wgpu::BufferUsages::COPY_DST,
+                mapped_at_creation: false,
+            });
+        let blurred_occlusion = ImageBuffer::new(
+            &self.gpu.device,
+            "ssao blurred occlusion".to_owned(),
+            ImageSize::new(image_size.width(), image_size.height()),
+        )?;
+        Ok(SsaoBuffers {
+            ssao_config,
+            blur_config,
+            raw_occlusion,
+            blurred_occlusion,
+        })
     }
 
     /// Submits a set of merge operations to combine all of the images
@@ -292,9 +372,9 @@ impl Context {
             }
         }
         buf.out
-            .grow_to_fit(&self.device, size)
+            .grow_to_fit(&self.gpu.device, size)
             .map_err(MergeError::OutputSize)?;
-        let mut encoder = self.device.create_command_encoder(
+        let mut encoder = self.gpu.device.create_command_encoder(
             &wgpu::CommandEncoderDescriptor { label: None },
         );
         // Scope to bound the lifetime of compute_pass
@@ -315,6 +395,7 @@ impl Context {
                 };
                 {
                     let mut writer = self
+                        .gpu
                         .queue
                         .write_buffer_with(
                             &buf.config,
@@ -334,8 +415,8 @@ impl Context {
                         .bind_active(),
                 };
 
-                let bg =
-                    self.device.create_bind_group(&wgpu::BindGroupDescriptor {
+                let bg = self.gpu.device.create_bind_group(
+                    &wgpu::BindGroupDescriptor {
                         label: Some("merge bind group"),
                         layout: &self.merge_bind_group_layout,
                         entries: &[
@@ -355,7 +436,8 @@ impl Context {
                                 resource: buf.out.bind_active(),
                             },
                         ],
-                    });
+                    },
+                );
                 compute_pass.set_bind_group(0, Some(&bg), &[]);
                 compute_pass.dispatch_workgroups(
                     size.width().div_ceil(8),
@@ -364,7 +446,7 @@ impl Context {
                 );
             }
         }
-        self.queue.submit(Some(encoder.finish()));
+        self.gpu.queue.submit(Some(encoder.finish()));
         Ok(())
     }
 
@@ -374,14 +456,15 @@ impl Context {
     pub fn submit_shade(
         &self,
         image: &MergeBuffers,
+        ssao: Option<&SsaoBuffers>,
         buf: &mut ShadeBuffers,
         out: Option<&mut ImageReadBuffer<ShadedImageTag>>,
     ) -> Result<(), ShadeError> {
         let size = image.out.size();
         buf.out
-            .grow_to_fit(&self.device, size)
+            .grow_to_fit(&self.gpu.device, size)
             .map_err(ShadeError::OutputSize)?;
-        let mut encoder = self.device.create_command_encoder(
+        let mut encoder = self.gpu.device.create_command_encoder(
             &wgpu::CommandEncoderDescriptor { label: None },
         );
 
@@ -395,10 +478,11 @@ impl Context {
             compute_pass.set_pipeline(&self.shade_pipeline);
             let cfg = ShadeConfig {
                 image_size: [size.width(), size.height(), image.depth],
-                _pad: 0u32,
+                has_ssao: ssao.is_some() as u32,
             };
             {
                 let mut writer = self
+                    .gpu
                     .queue
                     .write_buffer_with(
                         &buf.config,
@@ -409,24 +493,34 @@ impl Context {
                 writer.copy_from_slice(cfg.as_bytes());
             }
             let bg =
-                self.device.create_bind_group(&wgpu::BindGroupDescriptor {
-                    label: Some("shade bind group"),
-                    layout: &self.shade_bind_group_layout,
-                    entries: &[
-                        wgpu::BindGroupEntry {
-                            binding: 0,
-                            resource: buf.config.as_entire_binding(),
-                        },
-                        wgpu::BindGroupEntry {
-                            binding: 1,
-                            resource: image.out.bind_active(),
-                        },
-                        wgpu::BindGroupEntry {
-                            binding: 2,
-                            resource: buf.out.bind_active(),
-                        },
-                    ],
-                });
+                self.gpu
+                    .device
+                    .create_bind_group(&wgpu::BindGroupDescriptor {
+                        label: Some("shade bind group"),
+                        layout: &self.shade_bind_group_layout,
+                        entries: &[
+                            wgpu::BindGroupEntry {
+                                binding: 0,
+                                resource: buf.config.as_entire_binding(),
+                            },
+                            wgpu::BindGroupEntry {
+                                binding: 1,
+                                resource: image.out.bind_active(),
+                            },
+                            wgpu::BindGroupEntry {
+                                binding: 2,
+                                resource: ssao
+                                    .map(|s| {
+                                        s.blurred_occlusion().bind_active()
+                                    })
+                                    .unwrap_or_else(|| image.out.bind_active()),
+                            },
+                            wgpu::BindGroupEntry {
+                                binding: 3,
+                                resource: buf.out.bind_active(),
+                            },
+                        ],
+                    });
             compute_pass.set_bind_group(0, Some(&bg), &[]);
             compute_pass.dispatch_workgroups(
                 size.width().div_ceil(8),
@@ -435,7 +529,7 @@ impl Context {
             );
         }
         if let Some(image) = out {
-            image.grow_to_fit(&self.device, buf.out.size()).expect(
+            image.grow_to_fit(&self.gpu.device, buf.out.size()).expect(
                 "buf.out.size should always be \
                  a valid size for grow_to_fit",
             );
@@ -447,38 +541,347 @@ impl Context {
                 buf.out.size_bytes(),
             );
         }
-        self.queue.submit(Some(encoder.finish()));
+        self.gpu.queue.submit(Some(encoder.finish()));
         Ok(())
     }
 
-    /// Builds an [`ImageReadBuffer`]
-    ///
-    /// The output buffer is sized to read from a [`ShadeBuffers`] output
-    pub fn shaded_read_buffer(
+    /// Submits a pass to compute an SSAO buffer
+    pub fn submit_ssao(
         &self,
-        buffers: &ShadeBuffers,
-    ) -> ImageReadBuffer<ShadedImageTag> {
-        // TODO move this to new `Gpu` type?
-        ImageBuffer::new(
-            &self.device,
-            "shaded image".to_owned(),
-            buffers.out.size(),
-        )
-        .expect(
-            "buffers.out.size should always be \
-             a valid size for ImageBuffer::new",
-        )
-    }
-
-    /// Maps an image read buffer so that it can be read on the CPU
-    pub fn map_shaded_image<'a>(
-        &self,
-        image: &'a mut ImageReadBuffer<ShadedImageTag>,
-    ) -> MappedImage<'a, ShadedImageTag> {
-        // TODO move this to new `Gpu` type?
-        MappedImage::map(&self.device, image)
+        image: &MergeBuffers,
+        buf: &mut SsaoBuffers,
+    ) -> Result<(), SsaoError> {
+        self.ssao_ctx.submit(image, buf, &self.gpu)
     }
 }
+
+////////////////////////////////////////////////////////////////////////////////
+
+tag!(pub SsaoRawBufferTag, f32, STORAGE | COPY_SRC,
+    "Tag for a raw SSAO occlusion buffer");
+tag!(pub SsaoBlurredBufferTag, f32, STORAGE | COPY_SRC,
+    "Tag for a blurred SSAO occlusion buffer");
+
+/// Handle to a set of buffers used when running an SSAO pass
+pub struct SsaoBuffers {
+    ssao_config: wgpu::Buffer,
+    raw_occlusion: ImageBuffer<SsaoRawBufferTag>,
+
+    blur_config: wgpu::Buffer,
+    blurred_occlusion: ImageBuffer<SsaoBlurredBufferTag>,
+}
+
+impl SsaoBuffers {
+    /// Returns a handle to the raw SSAO occlusion buffer
+    pub fn raw_occlusion(&self) -> &ImageBuffer<SsaoRawBufferTag> {
+        &self.raw_occlusion
+    }
+
+    /// Returns a handle to the blurred SSAO occlusion buffer
+    pub fn blurred_occlusion(&self) -> &ImageBuffer<SsaoBlurredBufferTag> {
+        &self.blurred_occlusion
+    }
+}
+
+#[derive(Copy, Clone, FromBytes, Immutable, IntoBytes, KnownLayout)]
+#[repr(C)]
+struct SsaoConfig {
+    /// Image size, in voxels
+    image_size: [u32; 3],
+
+    /// Radius of SSAO sampling
+    radius: f32,
+}
+
+#[derive(Copy, Clone, FromBytes, Immutable, IntoBytes, KnownLayout)]
+#[repr(C)]
+struct BlurConfig {
+    /// Image size, in pixels
+    image_size: [u32; 2],
+
+    /// Pixel radius of blur
+    radius: u32,
+
+    /// Padding to 16 bytes
+    _pad: u32,
+}
+
+struct SsaoContext {
+    /// Fixed bind group for SSAO pass
+    ///
+    /// This contains the SSAO kernel and noise buffers, which are constants.
+    ssao_bind_group: wgpu::BindGroup,
+
+    /// Layout for bind group that accepts buffers from the user
+    ssao_bind_group_layout: wgpu::BindGroupLayout,
+
+    /// Pipeline for computing per-pixel SSAO
+    ssao_pipeline: wgpu::ComputePipeline,
+
+    /// Layout for blur pipeline
+    blur_bind_group_layout: wgpu::BindGroupLayout,
+
+    /// Pipeline for blurring an SSAO image
+    blur_pipeline: wgpu::ComputePipeline,
+}
+
+impl SsaoContext {
+    pub fn new(device: &wgpu::Device) -> Self {
+        let ssao_fixed_bind_group_layout =
+            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+                label: None,
+                entries: &[buffer_ro(0), buffer_ro(1)],
+            });
+        let ssao_user_bind_group_layout =
+            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+                label: None,
+                entries: &[buffer_uniform(0), buffer_ro(1), buffer_rw(2)],
+            });
+
+        const KERNEL_SIZE: usize = 64;
+        const NOISE_SIZE: usize = 16;
+
+        // Build constant buffers and their bind group
+        let ssao_kernel_size_bytes =
+            KERNEL_SIZE * std::mem::size_of::<[f32; 3]>();
+        let ssao_kernel = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("ssao kernel"),
+            size: ssao_kernel_size_bytes as u64,
+            usage: wgpu::BufferUsages::STORAGE,
+            mapped_at_creation: true,
+        });
+        let ssao_kernel_values =
+            fidget_raster::effects::ssao_kernel(KERNEL_SIZE);
+        ssao_kernel
+            .get_mapped_range_mut(0..ssao_kernel_size_bytes as u64)
+            .copy_from_slice(ssao_kernel_values.as_slice().as_bytes());
+        ssao_kernel.unmap();
+
+        let ssao_noise_size_bytes =
+            NOISE_SIZE * std::mem::size_of::<[f32; 2]>();
+        let ssao_noise = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("ssao noise"),
+            size: ssao_noise_size_bytes as u64,
+            usage: wgpu::BufferUsages::STORAGE,
+            mapped_at_creation: true,
+        });
+        let ssao_noise_values = fidget_raster::effects::ssao_noise(NOISE_SIZE);
+        ssao_noise
+            .get_mapped_range_mut(0..ssao_noise_size_bytes as u64)
+            .copy_from_slice(ssao_noise_values.as_slice().as_bytes());
+        ssao_noise.unmap();
+
+        let ssao_bind_group =
+            device.create_bind_group(&wgpu::BindGroupDescriptor {
+                label: Some("ssao fixed bind group"),
+                layout: &ssao_fixed_bind_group_layout,
+                entries: &[
+                    wgpu::BindGroupEntry {
+                        binding: 0,
+                        resource: ssao_kernel.as_entire_binding(),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 1,
+                        resource: ssao_noise.as_entire_binding(),
+                    },
+                ],
+            });
+
+        let shader_code = ssao_shader();
+        let pipeline_layout =
+            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+                label: Some("effects ssao pipeline"),
+                bind_group_layouts: &[
+                    Some(&ssao_user_bind_group_layout),
+                    Some(&ssao_fixed_bind_group_layout),
+                ],
+                immediate_size: 0u32,
+            });
+        let shader_module =
+            device.create_shader_module(wgpu::ShaderModuleDescriptor {
+                label: Some("effects ssao shader module"),
+                source: wgpu::ShaderSource::Wgsl(shader_code.into()),
+            });
+        let ssao_pipeline =
+            device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
+                label: Some("effects ssao compute pipeline"),
+                layout: Some(&pipeline_layout),
+                module: &shader_module,
+                entry_point: Some("ssao_main"),
+                compilation_options: Default::default(),
+                cache: None,
+            });
+
+        let blur_bind_group_layout =
+            device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+                label: None,
+                entries: &[buffer_uniform(0), buffer_ro(1), buffer_rw(2)],
+            });
+        let shader_code = blur_shader();
+        let pipeline_layout =
+            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+                label: Some("effects blur pipeline"),
+                bind_group_layouts: &[Some(&blur_bind_group_layout)],
+                immediate_size: 0u32,
+            });
+        let shader_module =
+            device.create_shader_module(wgpu::ShaderModuleDescriptor {
+                label: Some("effects blur shader module"),
+                source: wgpu::ShaderSource::Wgsl(shader_code.into()),
+            });
+        let blur_pipeline =
+            device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
+                label: Some("effects blur compute pipeline"),
+                layout: Some(&pipeline_layout),
+                module: &shader_module,
+                entry_point: Some("blur_main"),
+                compilation_options: Default::default(),
+                cache: None,
+            });
+
+        Self {
+            ssao_bind_group,
+            ssao_bind_group_layout: ssao_user_bind_group_layout,
+            ssao_pipeline,
+            blur_bind_group_layout,
+            blur_pipeline,
+        }
+    }
+
+    fn submit(
+        &self,
+        image: &MergeBuffers,
+        buf: &mut SsaoBuffers,
+        gpu: &Gpu,
+    ) -> Result<(), SsaoError> {
+        let image_size = image.out.size();
+        buf.raw_occlusion
+            .grow_to_fit(&gpu.device, image_size)
+            .map_err(SsaoError::OutputSize)?;
+        buf.blurred_occlusion
+            .grow_to_fit(&gpu.device, image_size)
+            .map_err(SsaoError::OutputSize)?;
+
+        // TODO make this passed in?
+        let mut encoder = gpu.device.create_command_encoder(
+            &wgpu::CommandEncoderDescriptor { label: None },
+        );
+
+        // Scope to bound the lifetime of compute_pass
+        {
+            let mut compute_pass =
+                encoder.begin_compute_pass(&wgpu::ComputePassDescriptor {
+                    label: None,
+                    timestamp_writes: None, // TODO add timestamps?
+                });
+            compute_pass.set_pipeline(&self.ssao_pipeline);
+            let cfg = SsaoConfig {
+                image_size: [
+                    image_size.width(),
+                    image_size.height(),
+                    image.depth,
+                ],
+                radius: 0.1,
+            };
+            {
+                let mut writer = gpu
+                    .queue
+                    .write_buffer_with(
+                        &buf.ssao_config,
+                        0,
+                        (std::mem::size_of::<SsaoConfig>() as u64)
+                            .try_into()
+                            .unwrap(),
+                    )
+                    .unwrap();
+                writer.copy_from_slice(cfg.as_bytes());
+            }
+
+            let bg = gpu.device.create_bind_group(&wgpu::BindGroupDescriptor {
+                label: Some("ssao bind group"),
+                layout: &self.ssao_bind_group_layout,
+                entries: &[
+                    wgpu::BindGroupEntry {
+                        binding: 0,
+                        resource: buf.ssao_config.as_entire_binding(),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 1,
+                        resource: image.out.bind_active(),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 2,
+                        resource: buf.raw_occlusion.bind_active(),
+                    },
+                ],
+            });
+            compute_pass.set_bind_group(0, Some(&bg), &[]);
+            compute_pass.set_bind_group(1, Some(&self.ssao_bind_group), &[]);
+            compute_pass.dispatch_workgroups(
+                image_size.width().div_ceil(8),
+                image_size.height().div_ceil(8),
+                1,
+            );
+        }
+
+        {
+            let mut compute_pass =
+                encoder.begin_compute_pass(&wgpu::ComputePassDescriptor {
+                    label: None,
+                    timestamp_writes: None, // TODO add timestamps?
+                });
+            compute_pass.set_pipeline(&self.blur_pipeline);
+            let cfg = BlurConfig {
+                image_size: [image_size.width(), image_size.height()],
+                radius: 2,
+                _pad: 0,
+            };
+            {
+                let mut writer = gpu
+                    .queue
+                    .write_buffer_with(
+                        &buf.blur_config,
+                        0,
+                        (std::mem::size_of::<BlurConfig>() as u64)
+                            .try_into()
+                            .unwrap(),
+                    )
+                    .unwrap();
+                writer.copy_from_slice(cfg.as_bytes());
+            }
+
+            let bg = gpu.device.create_bind_group(&wgpu::BindGroupDescriptor {
+                label: Some("blur bind group"),
+                layout: &self.blur_bind_group_layout,
+                entries: &[
+                    wgpu::BindGroupEntry {
+                        binding: 0,
+                        resource: buf.blur_config.as_entire_binding(),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 1,
+                        resource: buf.raw_occlusion.bind_active(),
+                    },
+                    wgpu::BindGroupEntry {
+                        binding: 2,
+                        resource: buf.blurred_occlusion.bind_active(),
+                    },
+                ],
+            });
+            compute_pass.set_bind_group(0, Some(&bg), &[]);
+            compute_pass.dispatch_workgroups(
+                image_size.width().div_ceil(8),
+                image_size.height().div_ceil(8),
+                1,
+            );
+        }
+
+        gpu.queue.submit(Some(encoder.finish()));
+        Ok(())
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////
 
 #[cfg(test)]
 mod test {
@@ -492,9 +895,12 @@ mod test {
     #[test]
     fn compile_shaders() {
         #[allow(clippy::single_element_loop)] // there will be more
-        for (src, desc) in
-            [(merge_shader(), "merge"), (shade_shader(), "shade")]
-        {
+        for (src, desc) in [
+            (merge_shader(), "merge"),
+            (shade_shader(), "shade"),
+            (ssao_shader(), "ssao"),
+            (blur_shader(), "blur"),
+        ] {
             crate::compile_shader(&src, desc);
         }
     }