+++ fidget-wgpu/src/lib.rs
@@ -2,6 +2,7 @@
 #![warn(missing_docs)]
 
 use heck::ToShoutySnakeCase;
+use zerocopy::{FromBytes, Immutable};
 
 pub mod buf;
 pub mod effects;
@@ -25,7 +26,7 @@ fn opcode_constants() -> String {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-/// Error type for [`init`]
+/// Error type for GPU initialization
 #[derive(Debug, thiserror::Error)]
 pub enum InitError {
     /// Error when requesting an adapter
@@ -37,29 +38,108 @@ pub enum InitError {
     Device(#[from] wgpu::RequestDeviceError),
 }
 
-/// Returns a WebGPU device and queue with appropriate settings
-///
-/// Non-default settings are as follows:
-/// - We request a [`wgpu::PowerPreference::HighPerformance`] adapter
-/// - We enable the [`wgpu::Features::TIMESTAMP_QUERY`] feature
-///
-/// This is a helper function for simplicity; more sophisticated systems will
-/// likely construct the adapter, device, and queue themselves.
-pub async fn init() -> Result<(wgpu::Device, wgpu::Queue), InitError> {
-    let instance = wgpu::Instance::default();
-    let adapter = instance
-        .request_adapter(&wgpu::RequestAdapterOptions {
-            power_preference: wgpu::PowerPreference::HighPerformance,
-            ..wgpu::RequestAdapterOptions::default()
-        })
-        .await?;
-    let out = adapter
-        .request_device(&wgpu::DeviceDescriptor {
-            required_features: wgpu::Features::TIMESTAMP_QUERY,
-            ..wgpu::DeviceDescriptor::default()
-        })
-        .await?;
-    Ok(out)
+/// Handle to a GPU device
+#[derive(Clone)]
+pub struct Gpu {
+    /// GPU device
+    pub device: wgpu::Device,
+    /// GPU queue
+    pub queue: wgpu::Queue,
+}
+
+impl Gpu {
+    /// Returns a [`Gpu`] object with customized settings
+    ///
+    /// Non-default settings are as follows:
+    /// - We request a [`wgpu::PowerPreference::HighPerformance`] adapter
+    /// - We enable the [`wgpu::Features::TIMESTAMP_QUERY`] feature
+    ///
+    /// This is a helper function for simplicity; more sophisticated systems
+    /// will likely construct the adapter, device, and queue themselves.
+    pub async fn init() -> Result<Gpu, InitError> {
+        let instance = wgpu::Instance::default();
+        let adapter = instance
+            .request_adapter(&wgpu::RequestAdapterOptions {
+                power_preference: wgpu::PowerPreference::HighPerformance,
+                ..wgpu::RequestAdapterOptions::default()
+            })
+            .await?;
+        let (device, queue) = adapter
+            .request_device(&wgpu::DeviceDescriptor {
+                required_features: wgpu::Features::TIMESTAMP_QUERY,
+                ..wgpu::DeviceDescriptor::default()
+            })
+            .await?;
+        Ok(Gpu { device, queue })
+    }
+
+    /// Returns a [`Gpu`] object with default settings
+    ///
+    /// This is useful for CI, where `TIMESTAMP_QUERY` is unsupported
+    #[doc(hidden)]
+    pub async fn init_basic() -> Result<Gpu, InitError> {
+        let instance = wgpu::Instance::default();
+        let adapter = instance
+            .request_adapter(&wgpu::RequestAdapterOptions::default())
+            .await?;
+        let (device, queue) = adapter
+            .request_device(&wgpu::DeviceDescriptor::default())
+            .await?;
+        Ok(Gpu { device, queue })
+    }
+
+    /// Returns a readable buffer for the given image buffer
+    pub fn read_buffer_for<T: buf::BufferTag>(
+        &self,
+        buf: &buf::ImageBuffer<T>,
+    ) -> buf::ImageReadBuffer<T> {
+        buf::ImageBuffer::new(
+            &self.device,
+            format!("{} (read)", buf.name()),
+            buf.size(),
+        )
+        .expect("buf.size should always be a valid size for ImageBuffer::new")
+    }
+
+    /// Maps a readable image buffer, returning a mapped image
+    pub fn map<'a, T: buf::BufferTag>(
+        &self,
+        buf: &'a mut buf::ImageReadBuffer<T>,
+    ) -> buf::MappedImage<'a, T> {
+        buf::MappedImage::map(&self.device, buf)
+    }
+
+    /// Debug function to read from a buffer to a `Vec<T>`
+    pub fn read_vec<T: FromBytes + Immutable + Clone + Copy>(
+        &self,
+        buf: &wgpu::Buffer,
+    ) -> Vec<T> {
+        let scratch = self.device.create_buffer(&wgpu::BufferDescriptor {
+            label: None,
+            size: buf.size(),
+            usage: wgpu::BufferUsages::COPY_DST | wgpu::BufferUsages::MAP_READ,
+            mapped_at_creation: false,
+        });
+        let mut encoder = self.device.create_command_encoder(
+            &wgpu::CommandEncoderDescriptor {
+                label: Some("read_buffer"),
+            },
+        );
+        encoder.copy_buffer_to_buffer(buf, 0, &scratch, 0, buf.size());
+        self.queue.submit(Some(encoder.finish()));
+
+        let buffer_slice = scratch.slice(..);
+        buffer_slice.map_async(wgpu::MapMode::Read, |_| {});
+        self.device
+            .poll(wgpu::PollType::wait_indefinitely())
+            .unwrap();
+
+        let result = <[T]>::ref_from_bytes(&buffer_slice.get_mapped_range())
+            .unwrap()
+            .to_vec();
+        scratch.unmap();
+        result
+    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -112,32 +192,26 @@ mod test {
             return;
         }
 
-        let instance = wgpu::Instance::default();
-        let (device, queue) = pollster::block_on(async {
-            let adapter = instance
-                .request_adapter(&wgpu::RequestAdapterOptions::default())
-                .await
-                .unwrap();
-            adapter
-                .request_device(&wgpu::DeviceDescriptor::default())
-                .await
-                .unwrap()
-        });
+        let gpu = pollster::block_on(Gpu::init_basic()).unwrap();
+        let voxel_ctx = voxel::Context::new(&gpu);
+        let effects_ctx = effects::Context::new(&gpu);
 
-        let voxel_ctx = voxel::Context::new(device.clone(), queue.clone());
-        let effects_ctx = effects::Context::new(device.clone(), queue.clone());
-
-        let size = 32;
+        let size = 128;
         let image_size = RenderSize::from(size);
         let mut buf = voxel_ctx.buffers(image_size).unwrap();
         let mut merge_buf = effects_ctx.merge_buffers(size.into()).unwrap();
         let mut shade_buf = effects_ctx.shade_buffers(size.into()).unwrap();
-        let mut shade_out = effects_ctx.shaded_read_buffer(&shade_buf);
+        let mut shade_out = gpu.read_buffer_for(shade_buf.output());
 
         let (x, y, z) = Tree::axes();
-        let sphere =
-            (x.square() + y.square() + z.square()).sqrt() - Tree::constant(0.5);
-        let shape = voxel_ctx.shape(&VmShape::from(sphere)).unwrap();
+        let x_ = x.clone() - 0.2;
+        let sphere1 = (x_.square() + y.square() + z.square()).sqrt()
+            - Tree::constant(0.5);
+        let x_ = x + 0.2;
+        let sphere2 = (x_.square() + y.square() + z.square()).sqrt()
+            - Tree::constant(0.5);
+        let spheres = sphere1.min(sphere2);
+        let shape = voxel_ctx.shape(&VmShape::from(spheres)).unwrap();
 
         voxel_ctx
             .submit(
@@ -152,12 +226,19 @@ mod test {
         effects_ctx
             .submit_merge(&[buf.image_storage_buffer()], true, &mut merge_buf)
             .unwrap();
+        let mut ssao_buf = effects_ctx.ssao_buffers(size.into()).unwrap();
+        effects_ctx.submit_ssao(&merge_buf, &mut ssao_buf).unwrap();
         effects_ctx
-            .submit_shade(&merge_buf, &mut shade_buf, Some(&mut shade_out))
+            .submit_shade(
+                &merge_buf,
+                Some(&ssao_buf),
+                &mut shade_buf,
+                Some(&mut shade_out),
+            )
             .unwrap();
-        let img = effects_ctx.map_shaded_image(&mut shade_out);
-        let (_out, size) = img.image().take();
-        assert_eq!(size.width(), 32);
-        assert_eq!(size.height(), 32);
+        let img = gpu.map(&mut shade_out);
+        let (_out, img_size) = img.image().take();
+        assert_eq!(img_size.width(), size);
+        assert_eq!(img_size.height(), size);
     }
 }